requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==1.2.0
filelock==3.12.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
pytest-testdox==3.1.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.1
tzdata==2023.3
virtualenv==20.24.5
wcwidth==0.2.6
